[testing] make sure tests cleanup their instances directories and temporary files (#18079)
* make sure the dump tests cleanup their instances directories

* cleanup at the end of the run.

* make sure server parameters tests clean up at the end on success.

* fix cleanup

* fix directory cleanup

* don't clean up if an error occured

* cleanup: we don't care if there is nothing to cleanup anymore.

* windows needs checks here

* different base directories!

* fix order of shutdown & cleanup

* fix launch sequence

* properly shutdown test as we do in the other cases

---------

+++ js/client/modules/@arangodb/testutils/instance.js
@@ -512,14 +512,21 @@ class instance {
     if (this.options.extremeVerbosity) {
       print(CYAN + "cleaning up " + this.name + " 's Directory: " + this.rootDir + RESET);
     }
-    fs.removeDirectoryRecursive(this.rootDir, true);
+    if (fs.exists(this.rootDir)) {
+      fs.removeDirectoryRecursive(this.rootDir, true);
+    }
   }
 
   // //////////////////////////////////////////////////////////////////////////////
   // / @brief scans the log files for assert lines
   // //////////////////////////////////////////////////////////////////////////////
 
-  readAssertLogLines () {
+  readAssertLogLines (expectAsserts) {
+    if (!fs.exists(this.logFile)) {
+      if (fs.exists(this.rootDir)) {
+        print(`readAssertLogLines: Logfile ${this.logFile} already gone.`);
+      }
+      return;
     let size = fs.size(this.logFile);
     if (this.options.maxLogFileSize !== 0 && size > this.options.maxLogFileSize) {
       // File bigger 500k? this needs to be a bug in the tests.
